requirements.txt
# Freepik ComfyUI Custom Node Requirements

# Core dependencies
requests>=2.31.0
Pillow>=10.0.0
torch>=2.0.0

# Optional dependencies for enhanced features
urllib3>=2.0.0

# Already included in ComfyUI
# numpy
# torchvision
